archive/llm_to_kg/requirements.txt
accelerate==1.11.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
attrs==25.4.0
beautifulsoup4==4.14.2
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
colorlog==6.10.1
contourpy==1.3.3
cycler==0.12.1
dill==0.4.0
docling==2.58.0
docling-core==2.49.0
docling-ibm-models==3.10.2
docling-parse==4.7.0
et_xmlfile==2.0.0
Faker==37.12.0
filelock==3.20.0
filetype==1.2.0
fonttools==4.60.1
fsspec==2025.9.0
hf-xet==1.2.0
huggingface-hub==0.36.0
idna==3.11
Jinja2==3.1.6
jsonlines==4.0.0
jsonref==1.1.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
latex2mathml==3.78.1
lxml==5.4.0
markdown-it-py==4.0.0
marko==2.2.1
matplotlib==3.10.7
mdurl==0.1.2
mpire==2.10.2
mpmath==1.3.0
multiprocess==0.70.18
networkx==3.5
numpy==2.2.6
omegaconf==2.3.0
opencv-python==4.12.0.88
openpyxl==3.1.5
packaging==25.0
pandas==2.3.3
pdf2image==1.17.0
pillow==11.3.0
polyfactory==2.22.3
psutil==7.1.2
pyclipper==1.3.0.post6
pydantic==2.12.3
pydantic_core==2.41.4
pydantic-settings==2.11.0
Pygments==2.19.2
pylatexenc==2.10
pyparsing==3.2.5
pypdfium2==4.30.0
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.2.1
python-pptx==1.0.2
pytz==2025.2
pywin32==311
PyYAML==6.0.3
rapidocr==3.4.2
referencing==0.37.0
regex==2025.10.23
requests==2.32.5
rich==14.2.0
rpds-py==0.28.0
rtree==1.4.1
safetensors==0.6.2
scipy==1.16.3
semchunk==2.2.2
shapely==2.1.2
shellingham==1.5.4
soupsieve==2.8
sympy==1.13.1
tabulate==0.9.0
tokenizers==0.22.1
torch==2.5.1+cu121
torchaudio==2.5.1+cu121
torchvision==0.20.1+cu121
tqdm==4.67.1
transformers==4.57.1
typer==0.19.2
typing-inspection==0.4.2
tzdata==2025.2
urllib3==2.5.0
xlsxwriter==3.2.9
